@@ -1,6 +1,6 @@
 package net.javaguides.springboot.tutorial.controller;
 
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
